requirements.txt
# requirements.txt
PyPDF2
python-docx